// projects/facility-management/src/app/core/enums/menu-licensing.enum.ts
/** menu-licensing index */
export enum MenuLicensing {
    Dashboard = 1,
    CopyIt = 2,
    BookIt = 3,
    Fleet = 4,
    VisitorLog = 5,
    Report = 6,
    ReportCopyIt = 7,
    ReportBookIt = 8,
    ReportFleet = 9,
    ReportMail = 10,
    ReportWorkflow = 11,
    ReportTask = 12,
    Packages = 13,
    ReportMeterReads = 14
}

/** available features for licensing */
export enum LicenseFeature {
    bookIt = 'bookIt',
    copyIt = 'copyIt',
    mail = 'mail',
    workflow = 'workflow',
    visitorLog = 'visitorLog',
}